fix(posix): set WEBI_CHECKSUM in both bootstrap and installer
